+++ packages/video/CHANGELOG.md
@@ -1,5 +1,11 @@
 # video
 
+## 0.1.15
+
+### Patch Changes
+
+- fix
+
 ## 0.1.14
 
 ### Patch Changes